Ancestral State Estimation based on Marginal Reconstruction for the MiSSE models.

Estimates the likeliest states for both internal nodes and tips of a phylogeny using the marginal reconstruction algorithm.

MarginReconMiSSE(phy, f, pars, hidden.states=1, fixed.eps=NULL,
condition.on.survival=TRUE,  root.type="madfitz", root.p=NULL, includes.fossils=FALSE, 
k.samples=NULL, strat.intervals=NULL, AIC=NULL, get.tips.only=FALSE, verbose=TRUE, 
n.cores=NULL, dt.threads=1)

Arguments

phy

a phylogenetic tree, in ape “phylo” format and with internal nodes labeled denoting the ancestral selective regimes.

f

the estimated proportion of extant species included in the phylogeny. A value of 0.50 means that 50 percent of species are contained in the. By default all species are assumed to be sampled.

pars

vector containing the MLE of the parameters.

hidden.states

a numeric indicating the number of shifts. The default is 1 meaning a single birth-death model.

fixed.eps

a value to be used to fix extinction fraction during search. Default is NULL meaning that it is freely estimated.

condition.on.survival

a logical indicating whether the likelihood should be conditioned on the survival of two lineages and the speciation event subtending them (Nee et al. 1994). The default is TRUE.

root.type

indicates whether root prior assumption should based the procedure described by FitzJohn et al. 2009, “madfitz”, assumed equal, “equal”, or set to user, “user”.

root.p

a vector indicating fixed root state probabilities. The default is NULL.

includes.fossils

a logical indicating whether the tree contains fossil taxa. The default is FALSE.

k.samples

a table of extinct individuals with sampled descendants. See details for how the table must be formatted.

strat.intervals

a table of extinct individuals with sampled descendants. See vignette for how the table must be formatted.

AIC

the AIC for the model being used for the reconstruction. This is used by the plotting function. The default is NULL.

verbose

a logical indicating whether progress should be printed to screen. The default is TRUE.

get.tips.only

a logical indicating whether just tip reconstructions should be output. The default is FALSE.

n.cores

specifies the number of independent processors to conduct the analysis. The default is NULL.

dt.threads

sets the number of threads available to data.table. In practice this need not change from the default of 1 thread, as we have not seen any speedup from allowing more threads.

Details

In this implementation the marginal probability of state combination i for a focal node is simply the overall likelihood of the tree and data when the state of the focal node is fixed in state combination i.

See help for “MiSSE” for more on other parameters for this function.

Value

MarginReconMiSSE returns an object of class misse.states. This is a list with elements:

$node.mat

the marginal probabilities calculated for each node. They are ordered based on the elements in the edge matrix in “phylo” format.

$tip.mat

the marginal probabilities calculated for each tip. They are ordered based on the order of tip labels in the tree.

$rate.mat

a matrix that details the rates for each state combination. This is used by the plotting function.

$phy

a phylogenetic tree in the “phylo” format that contains the states with the highest marginal probability at each internal node.

References

Beaulieu, J.M, and B.C. O'Meara. 2016. Detecting hidden diversification shifts in models of trait-dependent speciation and extinction. Syst. Biol. 65:583-601.

FitzJohn R.G., Maddison W.P., and Otto S.P. 2009. Estimating trait-dependent speciation and extinction rates from incompletely resolved phylogenies. Syst. Biol. 58:595-611.

Maddison W.P., Midford P.E., and Otto S.P. 2007. Estimating a binary characters' effect on speciation and extinction. Syst. Biol. 56:701-710.

Nee S., May R.M., and Harvey P.H. 1994. The reconstructed evolutionary process. Philos. Trans. R. Soc. Lond. B Biol. Sci. 344:305-311.

Author

Jeremy M. Beaulieu