Transition Rate matrix generator for the GeoHiSSE model

Generates and manipulates the index of the transition rate parameters to be optimized.

TransMatMakerGeoHiSSE(hidden.traits=0, make.null=FALSE, include.jumps=FALSE, 
  separate.extirpation=FALSE)

Arguments

hidden.traits

a numeric value with the number of hidden trait in the model. The cannonical GeoSSE model has no hidden areas, so 'hidden.traits=0'. The default value is 0.

make.null

Sets the transition matrix to the null model such that (1) transition rates from endemic to widespread areas (A -> AB and B -> AB) are the same across all hidden areas and (2) the transition rates between hidden areas (such as 0A <-> 1A) are the same for all areas.

include.jumps

allows for 0->1 and 1->0 transitions.

separate.extirpation

allows for 01->0 and 01->1 transitions.

Details

Outputs the full index of the rate parameters that are to be optimized.

The intention is that a user might want to see how the matrix is designed prior to an analysis and perhaps drop a few parameters beforehand due to some hypothesis that she might have. The resulting matrix is to be plugged directly into GeoHiSSE.

Value

Returns a rate matrix index.